Move NodePoolConfig to config.go
