--- core/chains/evm/config/config_node_pool.go
@@ -1,10 +1,0 @@
-package config
-
-import "time"
-
-type NodePool interface {
-	PollFailureThreshold() uint32
-	PollInterval() time.Duration
-	SelectionMode() string
-	SyncThreshold() uint32
-}
