+++ core/chains/evm/config/config_test.go
@@ -448,4 +448,20 @@ func Test_chainScopedConfig_Validate(t *testing.T) {
 	})
 }
 
+func TestNodePoolConfig(t *testing.T) {
+	gcfg := configtest.NewGeneralConfig(t, func(c *chainlink.Config, s *chainlink.Secrets) {
+		id := utils.NewBig(big.NewInt(rand.Int63()))
+		c.EVM[0] = &v2.EVMConfig{
+			ChainID: id,
+			Chain:   v2.Defaults(id, &v2.Chain{}),
+		}
+	})
+	cfg := evmtest.NewChainScopedConfig(t, gcfg)
+
+	require.Equal(t, "HighestHead", cfg.EVM().NodePool().SelectionMode())
+	require.Equal(t, uint32(5), cfg.EVM().NodePool().SyncThreshold())
+	require.Equal(t, time.Duration(10000000000), cfg.EVM().NodePool().PollInterval())
+	require.Equal(t, uint32(5), cfg.EVM().NodePool().PollFailureThreshold())
+}
+
 func ptr[T any](t T) *T { return &t }
